+++ contracts/IVoucherHub.sol
@@ -4,7 +4,16 @@
 pragma solidity ^0.8.20;
 
 interface IVoucherHub {
-    event VoucherCreated(address indexed voucher, address owner, uint256 expiration);
+    struct VoucherType {
+        string description;
+        uint256 duration;
+    }
+    event VoucherCreated(
+        address indexed voucher,
+        address owner,
+        uint256 voucherType,
+        uint256 expiration
+    );
     event VoucherTypeCreated(uint256 indexed id, string description, uint256 duration);
     event VoucherTypeDescriptionUpdated(uint256 indexed id, string description);
     event VoucherTypeDurationUpdated(uint256 indexed id, uint256 duration);
@@ -12,8 +21,26 @@ interface IVoucherHub {
     event EligibleAssetRemoved(uint256 indexed id, address asset);
 
     function createVoucher(
-        address account,
-        uint256 expiration
+        address owner,
+        uint256 voucherType
     ) external returns (address voucherAddress);
-    function getVoucher(address account) external view returns (address voucherAddress);
+
+    function getVoucher(address account) external view returns (address);
+
+    function getVoucherBeacon() external view returns (address);
+
+    function getIexecPoco() external view returns (address);
+
+    function getVoucherTypeCount() external view returns (uint256);
+
+    function getVoucherType(uint256 id) external view returns (VoucherType memory);
+
+    function addEligibleAsset(uint256 voucherTypeId, address asset) external;
+
+    function removeEligibleAsset(uint256 voucherTypeId, address asset) external;
+
+    function isAssetEligibleToMatchOrdersSponsoring(
+        uint256 voucherTypeId,
+        address asset
+    ) external view returns (bool);
 }