+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Create voucher from VoucherHub with : type, expiration, authorize list. (#6)
 - Create vouchers with create2. (#5)
 - Create upgradeable voucher contract. (#4)
 - Add voucher type structure, duration, description and asset eligible. (#3)